+++ entrypoint/core.py
@@ -1,11 +1,13 @@
-from typing import Callable, Type, TypeVar, overload
+from typing import Optional, TypeVar, final
 
 from attrs import frozen
+from named import get_module
+from typing_aliases import Nullary
 
 __all__ = ("MAIN", "Main", "EntryPoint", "entrypoint", "is_main")
 
-Main = Callable[[], None]
-"""The `main` function that does not return anything."""
+Main = Nullary[None]
+"""The `main` function that does not take arguments and returns nothing."""
 
 M = TypeVar("M", bound=Main)
 
@@ -25,36 +27,29 @@ def is_main(name: str) -> bool:
     return name == MAIN
 
 
+@final
 @frozen()
 class EntryPoint:
     """Represents handlers for [`@entrypoint`][entrypoint.core.entrypoint] decorators."""
 
-    name: str
-
-    def call(self, main: M) -> M:
-        if is_main(self.name):
-            main()
-
-        return main
+    name: Optional[str] = None
 
     def __call__(self, main: M) -> M:
         return self.call(main)
 
+    def call(self, main: M) -> M:
+        name = self.name
 
-EP = TypeVar("EP", bound=EntryPoint)
+        if name is None:
+            name = get_module(main)
 
+        if is_main(name):
+            main()
 
-@overload
-def entrypoint(name: str) -> EntryPoint:
-    ...
-
-
-@overload
-def entrypoint(name: str, entrypoint_type: Type[EP]) -> EP:
-    ...
+        return main
 
 
-def entrypoint(name: str, entrypoint_type: Type[EntryPoint] = EntryPoint) -> EntryPoint:
+def entrypoint(name: Optional[str] = None) -> EntryPoint:
     """Defines decorated functions as entry points.
 
     Calls the wrapped function if the module gets run directly.
@@ -63,15 +58,16 @@ def entrypoint(name: str, entrypoint_type: Type[EntryPoint] = EntryPoint) -> Ent
     callers to pass the `__name__` variable as an argument,
     and merely checks it against `__main__` when needed.
 
-    `entrypoint_type(name)` is created under the hood, and is
+    In case `__name__` is not provided, this function will attempt to read
+    the module name from the function given (via `__module__`).
+
+    `EntryPoint(name)` is created under the hood, and is
     then used to handle calls.
 
     Args:
         name: The `__name__` of the module.
-        entrypoint_type: The [`EntryPoint`][entrypoint.core.EntryPoint]
-            type that is used to handle calls.
 
     Returns:
         The created [`EntryPoint`][entrypoint.core.EntryPoint] instance.
     """
-    return entrypoint_type(name)
+    return EntryPoint(name)